parse json in amp-story
