+++ examples/amp-story/shopping.html
@@ -64,6 +64,7 @@
             },
             {
               "id": 4,
+              "tag-id": "banana-tag",
               "img": "/static/samples/img/product4_640x426.jpg",
               "name": "Banana",
               "attribution": "pixabay",
@@ -74,6 +75,7 @@
             },
             {
               "id": 5,
+              "tag-id": "watermelon-tag",
               "img": "/static/samples/img/product5_640x408.jpg",
               "name": "Watermelon",
               "attribution": "pixabay",
@@ -84,6 +86,7 @@
             },
             {
               "id": 6,
+              "tag-id": "melon-tag",
               "img": "/static/samples/img/product6_640x424.jpg",
               "name": "Melon",
               "attribution": "pixabay",
